change `get firmware` to `get component --firmware-only`
